requirements.txt
pandas
numpy
psycopg2-binary
openpyxl